Grocify-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "redis>=5.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.34.0",
]
